requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
gdown==5.2.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
# nvidia-cublas-cu11==11.11.3.6
# nvidia-cuda-cupti-cu11==11.8.87
# nvidia-cuda-nvrtc-cu11==11.8.89
# nvidia-cuda-runtime-cu11==11.8.89
# nvidia-cudnn-cu11==8.7.0.84
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.3.0.86
# nvidia-cusolver-cu11==11.4.1.48
# nvidia-cusparse-cu11==11.7.5.86
# nvidia-nccl-cu11==2.19.3
# nvidia-nvtx-cu11==11.8.86
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==10.2.0
psutil==7.0.0
py-cpuinfo==9.0.0
py_vncorenlp==0.1.4
pyjnius==1.6.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.13.1
six==1.17.0
soupsieve==2.7
sympy==1.13.3
tokenizers==0.21.2
# torch==2.2.2+cu118
# torchaudio==2.2.2+cu118
# torchvision==0.17.2+cu118
tqdm==4.67.1
transformers==4.53.0
triton==2.2.0
typing_extensions==4.9.0
tzdata==2025.2
ultralytics==8.3.160
ultralytics-thop==2.0.14
urllib3==2.5.0
vncorenlp==1.0.3
zipp==3.23.0
